[scanner] refactor: extract mission preview modal from FlightPlanBlueprint.tsx
